refactor(db): split admin dog profile query internals
- extract admin dog profile Prisma select shape into an internal module
- extract DB row mapping and profile types into dedicated files to reduce file size and improve maintainability

Refs: BEJ-110
