+++ packages/db/admin/dogs/profile/index.ts
@@ -1,4 +1,2 @@
-export {
-  getAdminDogProfileDb,
-  type AdminDogProfileDb,
-} from "./get-admin-dog-profile";
+export { getAdminDogProfileDb } from "./get-admin-dog-profile";
+export type { AdminDogProfileDb } from "./types";
